Added example for styled output.
This example needs some revision, together with the code in
`compare_timeseries_ref_data`. We pass the model data in to many different
methods over again, which both should be unnecessary, and also forces us to
recalculate from the beginning for every method call if we don't cache the
results (for the most part, it doesn't take long, but it smells). We should
either create a result object from the initial comparison, which we then pass
around to get different aspects of the  results outputted in various ways. Or
we should pass the output from one method on to the next, rather than the
IamDataFrame with the initial model data.

This commit also adds a temporary comment to compare_timeseries_ref_data that
touches on this point.